% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedxing.R
\name{foresee_next}
\alias{foresee_next}
\title{Foresee a next state}
\usage{
foresee_next(x0, v, pt, t0 = 0)
}
\arguments{
\item{x0}{an initial (pedestrian) position}

\item{v}{pedestrian walking speed}

\item{pt}{how far in time a pedestrian can foresee what consequence 
its action may result}

\item{t0}{how much time is needed to enact an action; assuming
instantaneous when t0 = 0; this should map to t0.}
}
\description{
Get a predictive next position for a pedestrian (R function is used 
in testing. Another copy of C++ function runs the model fitting.)
}
\details{
Function to get the predictive pedestrian state if beginning to cross. 
I assume a pedestrian prepares to walk to the north on the negative
side of y axis.
}
\examples{
## some examples
}
