% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get_traj}
\alias{get_traj}
\title{Get trajectory data out of a rModel function}
\usage{
get_traj(x, model = "inhibition")
}
\arguments{
\item{x}{a model output}

\item{model}{a string identify which model}
}
\description{
Get trajectory data out of a rModel function
}
\examples{
## some examples
}
