% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{Inhibition-class}
\alias{Inhibition-class}
\title{An S4 class for the Inhibition model}
\description{
An S4 class for the Inhibition model
}
\section{Slots}{

\describe{
\item{\code{choice_RT}}{choice_RT}

\item{\code{activation}}{activation}

\item{\code{dx}}{dx}

\item{\code{leak}}{leakage}

\item{\code{counter}}{counter}

\item{\code{samples}}{simulation samples}
}}

