% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{Input-class}
\alias{Input-class}
\title{An S4 class for the Input-only model}
\description{
An S4 class for the Input-only model
}
\section{Slots}{

\describe{
\item{\code{choice_RT}}{choice_RT}

\item{\code{activation}}{activation}

\item{\code{dx}}{dx}

\item{\code{counter}}{counter}
}}

