% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{LCA-class}
\alias{LCA-class}
\title{An S4 class}
\description{
An S4 class
}
\section{Slots}{

\describe{
\item{\code{choice_RT}}{choice_RT}

\item{\code{activation}}{activation}

\item{\code{inhibition}}{inhibition}

\item{\code{leakage}}{leakage}

\item{\code{dx}}{dx}

\item{\code{counter}}{counter}
}}

