% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_hist}
\alias{plot_hist}
\title{Plot choice RT distributions}
\usage{
plot_hist(x, binwidth = 0.1, cols = c("RT", "R"), print.fig = TRUE)
}
\arguments{
\item{x}{a data frame}

\item{binwidth}{bin width}

\item{cols}{key columns}

\item{print.fig}{Boolean}
}
\description{
Plot choice RT distributions
}
\examples{
## some examples
}
