% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulator.R
\name{rInput_R}
\alias{rInput_R}
\title{LBA Process in an R function}
\usage{
rInput_R(
  n,
  par,
  I,
  x0,
  maxiter = 5001,
  nonlinear = FALSE,
  random = FALSE,
  debug = FALSE
)
}
\arguments{
\item{n}{number of trials}

\item{par}{par}

\item{I}{I}

\item{x0}{x0}

\item{maxiter}{maxiter}

\item{nonlinear}{nonlinear}

\item{random}{random}

\item{debug}{debug}
}
\description{
Reconstruct an LBA in LCA structure
}
\examples{
## some examples
}
