% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulator.R
\name{rLCA_R}
\alias{rLCA_R}
\title{LCA Process in an R function}
\usage{
rLCA_R(
  n,
  par,
  I,
  x0,
  maxiter = 5001,
  nonlinear = FALSE,
  random = FALSE,
  debug = FALSE
)
}
\arguments{
\item{n}{number of trials}

\item{par}{par}

\item{I}{I}

\item{x0}{x0}

\item{maxiter}{maxiter}

\item{nonlinear}{nonlinear}

\item{random}{random}

\item{debug}{debug}
}
\description{
LCA Process in an R function
}
\examples{
p.vector <- c(kappa = 3, beta = 3, Z = .2, t0 = .3, dtovertau = 0.0001,
tau = 10)
args(rLCA_R)
res <- rLCA_R(n = 1, par = p.vector, I = c(1.2, 1), x0 = c(0, 0))

nacc <- length(I)
d1 <- data.table(x = rep(1:nrow(res[[1]]$trace), nacc), 
             y1 = as.vector \%>\% r(res[[1]]$trace), 
             y2 = as.vector(res[[1]]$lateral_inhibition),
             y3 = as.vector(res[[1]]$leakage),
             y4 = as.vector(res[[1]]$dx),
             z = factor(rep(1:nacc, each = nrow(res[[1]]$trace))))


p1 <- ggplot() +
geom_line(data = d1, aes(x=x, y=y1, colour =z)) +
  geom_hline(yintercept = Z[1]) + xlab("Time step") + ylab("Activation") +
theme_bw(base_size = 20) +
theme(aspect.ratio=1)
p1

}
